Add files via upload
This is a simple drawing program implemented in C++. It includes functions to draw three basic shapes: rectangles, right-angled triangles, and horizontal lines. Users can choose which shape they want to draw by entering their choice (1 for rectangle, 2 for triangle, and 3 for horizontal line) through the command line interface. For rectangles, users input the width and height; for triangles, they input the height; and for horizontal lines, they input the length. The program then displays the chosen shape using asterisks (*) in the console. This project is ideal for beginners learning C++ and looking to practice basic programming concepts such as functions, loops, and user input handling.